# openapi/markflow-backend.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aeon MarkFlow API",
    "version": "0.1.0",
    "description": "Aeon MarkFlow backend API — leads, engagement workflows, discovery calls."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
...
      "LeadResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "contact_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW-COLD",
              "NEW-INBOUND",
              "CONTACTED",
              "CONTACTED-PHONE",
              "CONTACTED-EMAIL",
              "PROSPECT",
              "INACTIVE",
              "RECLAIMED"
            ]
          },
          "email_deliverability": {
            "type": "string",
            "enum": [
              "GOOD",
              "LOW",
              "BAD"
            ]
          },
          "phone_dnd_status": {
            "type": "boolean"
          },
          "lost_reason": {
            "type": "string",
            "nullable": true
          },
          "lost_stage": {
            "type": "string",
            "nullable": true
          },
          "recycled_from_deal_id": {
            "type": "string",
            "nullable": true
          },
          "eligible_for_reengagement_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpsertGuardrailSettingsRequest": {
        "type": "object",
        "properties": {
          "ramp_up_starting_daily_cap": {
            "type": "number"
          },
          "ramp_up_step_multiplier": {
            "type": "number"
          },
          "ramp_up_step_interval_days": {
            "type": "number"
          },
          "ramp_up_steady_state_daily_cap": {
            "type": "number"
          },
          "guardrail_short_window_hours": {
            "type": "number"
          },
          "guardrail_long_window_days": {
            "type": "number"
          },
          "guardrail_min_sample_size": {
            "type": "number"
          },
          "throttle_bounce_rate": {
            "type": "number"
          },
          "throttle_complaint_rate": {
            "type": "number"
          },
          "hard_stop_bounce_rate": {
            "type": "number"
          },
          "hard_stop_complaint_rate": {
            "type": "number"
          }
        }
      },
      "GuardrailSettingsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UpsertGuardrailSettingsRequest"
          },
          {
            "type": "object",
            "properties": {
              "org_id": {
                "type": "string"
              },
              "domain": {
                "type": "string"
              },
              "has_override": {
                "type": "boolean"
              },
              "overridden_fields": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/orgs/{orgId}/leads": {
      "get": {
        "summary": "List/filter/search Leads for an org",
        "description": "Open to every role with org access, including BD-Lead Gen (view/upload/monitor only, per the RBAC table) — no workflow-related action happens here. `segment=recycled` narrows to the Recycled/Win-back segment (status RECLAIMED, with lost_reason/lost_stage populated) and is itself gated to BD-Sales/BD-Manager/Admin/Super Admin; any other role requesting it gets 403 rather than a silently-filtered result. `search` matches the lead's Contact name/company — since a Contact can be a Lead in more than one org, this always resolves back to *this org's own* Lead record, never another org's.\n",
        "tags": [
          "Leads"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "NEW-COLD",
                "NEW-INBOUND",
                "CONTACTED",
                "CONTACTED-PHONE",
                "CONTACTED-EMAIL",
                "PROSPECT",
                "INACTIVE",
                "RECLAIMED"
              ]
            }
          },
          {
            "in": "query",
            "name": "segment",
            "required": false,
            "description": "Pass \"recycled\" for the Recycled/Win-back segment (status RECLAIMED).",
            "schema": {
              "type": "string",
              "enum": [
                "recycled"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "description": "Case-insensitive substring match against the lead's Contact name/company.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "segment=recycled requested by a role outside RECYCLED_SEGMENT_ROLES"
          }
        }
      }
    },
    "/orgs/{orgId}/guardrail-settings": {
      "get": {
        "summary": "List SendGuardrail settings for every one of an org's configured sending domains",
        "description": "Admin/Super Admin only. One row per Organization.sending_domains[] entry, each showing the currently *effective* values (a per-domain override merged over the hardcoded defaults in src/constants/sendGuardrail.ts) plus which fields, if any, are overridden.\n",
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuardrailSettingsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{orgId}/guardrail-settings/{domain}": {
      "get": {
        "summary": "Get the effective SendGuardrail settings for one (org, domain) pair",
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK — always 200, even with no override (returns the hardcoded defaults)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardrailSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Upsert a per-(org, domain) override of SendGuardrail's ramp-up numbers/thresholds",
        "description": "Admin/Super Admin only. Only the fields present in the body are written — a field left out keeps its previously-stored value (or stays unset/default), the same partial-PATCH semantics as everywhere else in this codebase. Takes effect on the very next canSend() check; nothing needs a restart or redeploy.\n",
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertGuardrailSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardrailSettingsResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove the override — every field for this (org, domain) reverts to the hardcoded default",
        "description": "Admin/Super Admin only. Idempotent — deleting a pair with no override is a no-op.",
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "domain",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
...
}